+++ app/src/main/java/com/damsoft/overheidsdata/api/APIInterface.kt
@@ -3,8 +3,8 @@ package com.damsoft.overheidsdata.ui.api
 import android.arch.lifecycle.LiveData
 import com.damsoft.overheidsdata.api.ApiResponse
 import com.damsoft.overheidsdata.api.LiveDataCallAdapterFactory
-import com.damsoft.overheidsdata.ui.model.ArticlesResponse
-import com.damsoft.overheidsdata.ui.model.SourceResponse
+import com.damsoft.overheidsdata.model.ArticlesResponse
+import com.damsoft.overheidsdata.model.SourceResponse
 import retrofit2.Call
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
